[ROCm/gfx1250] Use x_fp8.dtype instead of hardcoded float8_e4m3fn in W4A8 GEMM1
GEMM1 output dtype was hardcoded to torch.float8_e4m3fn. If aiter's
downcast_to_static_fp8 returns a different FP8 type (e.g. e4m3fnuz on
a future ROCm build), the mismatch causes silent wrong results or a
type error in GEMM2.

Use x_fp8.dtype to match whatever type the downcast produced, making
the path arch-safe.

Mirrors ROCm/ATOM PR #1067 fix flagged in the Copilot review.

+++ python/sglang/srt/layers/moe/fused_moe_triton/aiter_mxfp4_w4a8_moe.py
@@ -184,7 +184,7 @@ def aiter_w4a8_gfx1250_forward(
         scatter_indx=None,
         gammas=None,
         swizzle_mx_scale=None,
-        out_dtype=torch.float8_e4m3fn,
+        out_dtype=x_fp8.dtype,
         apply_swiglu=True,
         alpha=gemm1_alpha,
         limit=gemm1_limit,
